Changed address bar text view maxHeight to CGFloat for convention.

+++ Code/LYRUIAddressBarTextView.m
@@ -13,7 +13,7 @@
 @interface LYRUIAddressBarTextView ()
 
 @property (nonatomic) UILabel *toLabel;
-@property (nonatomic) NSUInteger maxHeight;
+@property (nonatomic) CGFloat maxHeight;
 
 @end
